AP_NavEKF: Add option to use GPS height
Use of GPS alt is selected by setting the EKF_ALT_SOURCE parameter to 2
Reverts to Baro if GPS times out.
Baro height is used before a GPS origin is set.
An offset is calculated which is applied to Baro height to prevent height jumps when reverting, either due to to loss of GPS or changing the  parameter.
Baro height is used prior to setting the GPS origin.
Vertical accel aliasing checks are disabled if GPS height is being used because vertical position and velocity measurement errors are correlated
Report vertical position solution status unhealthy until GPS checks passed if using GPS height